+++ CHANGELOG.md
@@ -5,6 +5,10 @@ For releases after 1.3.0, the format is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Table based UTF-8 translation no longer mangles characters. [#14]
+
 ## [1.3.1] - 2021-01-30
 ### Fixed
 - Merged fix for Debian #861537, written by Vasily Kolobkov, passed on by